requirements.txt
redis==4.5.5
django==4.2.9
loguru==0.7.2
celery==5.3.6
Pillow==9.5.0
boto3==1.28.57
psycopg2-binary==2.9.9
whitenoise==6.4.0
python-decouple==3.8
drf-yasg==1.21.7
django-storages==1.14.2
django-celery-beat==2.5.0
django-import-export==3.2.0
djangorestframework==3.14.0
django-celery-results==2.5.1
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
django-filter==21.1
firebase-admin==6.4.0
celery==5.3.6
flower==2.0.1
django-celery-beat==2.5.0
django-redis==5.4.0
supabase==2.4.6
setuptools==75.1.0


